chore: optimize the integration test UI
