+++ frontend/appflowy_flutter/integration_test/cover_image_test.dart
@@ -29,7 +29,8 @@ void main() {
 
       await tester.tapGoButton();
       await tester.hoverOnCoverPluginAddButton();
-      await tester.expectToSeePluginAddCoverAndIconButton();
+
+      tester.expectToSeePluginAddCoverAndIconButton();
     });
   });
 }
