+++ frontend/appflowy_flutter/integration_test/util/util.dart
@@ -2,3 +2,4 @@ export 'base.dart';
 export 'common_operations.dart';
 export 'settings.dart';
 export 'data.dart';
+export 'expectation.dart';
